1. poisson and young estimator now works on a fractional part of the speciment 2. do not switch ui automatically if selected explcitly 3. body::maskOK to check mask

+++ core/NullGUI.cpp
@@ -114,7 +114,7 @@ int NullGUI::run(int argc, char* argv[])
 	} 
 	else // FileGenerator
 	{
-		for(unsigned int i=0; i<inputFiles.size() ; ++i)
+		for(size_t i=0; i<inputFiles.size() ; ++i)
 		{
 			std::cerr << "filegenerator: \"" << inputFiles[i] << "\"";
 			file=inputFiles[i];
